requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
sentence-transformers==3.3.1
faiss-cpu==1.13.2
google-genai==2.0.1
python-dotenv==1.0.1
httpx==0.27.2
